lesson-3: rename packages
