+++ app/src/main/java/com/rogoz208/nasapicture/ui/screens/pod/NasaPodViewModel.kt
@@ -1,4 +1,4 @@
-package com.rogoz208.nasapicture.ui.screens.picture
+package com.rogoz208.nasapicture.ui.screens.pod
 
 import android.util.Log
 import androidx.lifecycle.*
