+++ app/src/main/res/layout/fragment_nasa_pod.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.rogoz208.nasapicture.ui.screens.picture.NasaPodFragment">
+    tools:context="com.rogoz208.nasapicture.ui.screens.pod.NasaPodFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
